0.12.x `testEnableWorkStealing` cleanups (#4689)
`Enable` is a bit redundant since it's already a `Boolean`,
`WorkStealing` is misleading, and `Parallelism` is an important word to
get in there since this flag controls whether tests run serially or in
parallel. This flag will probably be the default mechanism for turning
on parallelism, over `testForkGrouping`, so it deserves the short name

In general replaced a lot of the "steal" terminology with "queue"
terminology

The logic pulling test class marker files out of the queue folder could
also be significantly simplified.



Also cut the test worker polling interval from 1000ms to 20s to try and
reduce the noticeable delay after tests start running before the prompt
updates