Support `mill init` from an existing sbt project, and improve related Maven and Gradle support (#4586)
Resolves #3450.

## Feature summary

The conversion
- handles deeply nested modules
- captures publish settings
- configures dependencies for configurations:
  - no configuration
  - Compile
  - Test
  - Runtime
  - Provided
  - Optional
- configures testing frameworks:
  - Java:
    - JUnit 4
    - JUnit 5
    - TestNG
  - Scala:
    - ScalaTest
    - Specs2
    - µTest
    - MUnit
    - Weaver
    - ZIOTest
    - ScalaCheck

The conversion does not support:
- custom dependency configurations
- custom sbt settings including custom tasks
- sources other than Scala on JVM and Java, such as Scala.js and Scala
Native
- cross builds

## Implementation summary

To resolve the issue, I use sbt's [addPluginSbtFile
command](https://www.scala-sbt.org/1.x/docs/Combined+Pages.html#addPluginSbtFile+command)
on Linux or add a temporary `.sbt` script in the sbt `project` directory
on Windows, to add a plugin jar assembled from the `exportplugin`
module, which adds an sbt task `millInitExportBuild` to extract the
settings for a build and serialize them with uPickle, and then the rest
of `SbtBuildGenMain` imports the file and generates the build. The
`exportplugin` sbt plugin module depends on sbt which requires Scala
2.12, so the `models` module is set up as a cross module.

Code in the common `buildgen`, Maven, and Gradle modules is refactored
and improved. Major changes include:

- The project hierarchy representation is updated from `Tree[Node[T]]`
to `Tree[Node[Option[T]]]` and related code is refactored.
- Some general integration test helpers such as `testMillInit` are added
in `MillInitUtils`, and the integration tests for Maven and Gradle and
updated using them.
- Generated module settings that are different from those in the base
module are overridden, and extended from their super ones if possible.
- Update some nullable types with Scala 3 explicit nulls.
- Inter-project dependencies for Gradle and sbt are imported in an
improved direct way rather than resolved from a GAV-module map, so
dependencies on subprojects with the same name are properly imported.

## Notes for further development

* I left some TODO comments in the code related to alternative
implementations, unimplemented features I don't know whether are
necessary at the moment and may involve some work, and
refactoring/renaming suggestions.

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>
Co-authored-by: Tobias Roeser <le.petit.fou@web.de>
Co-authored-by: Li Haoyi <haoyi.sg@gmail.com>