Replace uppercase "SBT" with lowercase "sbt", and update some links to `https` BTW (#4706)
The lowercase spelling "sbt" seems to be the official spelling used in
their docs and other places. Shall we replace the uppercase "SBT"s with
this instead?

This is done for all occurences of "SBT" except for the link
http://www.lihaoyi.com/post/SowhatswrongwithSBT.html which will be
broken if updated.

Pull Request: https://github.com/com-lihaoyi/mill/pull/4706
---------

Co-authored-by: Tobias Roeser <le.petit.fou@web.de>