Ignore some example tests that consistently fail on CI
This makes some example test failures be ignored on CI. The examples are
still run, their stacktrace when they fail is still printed, but this
doesn't make the CI job fail.

Whether the errors of an example tests are ignored or not is controled
by an `ignoreErrorsOnCI` file at its root. If this file exists, then
the test errors are printed but ignored. Only so on CI, not locally.