Remove dist/ from .gitignore
It seems some files used to be written there when running `./mill -i dist.installLocal`
I think, but that's not the case anymore (unless I'm mistaken?). So there's no need to
keep that entry there.

Having it there makes the global search of VSCode ignore that directory and its
`package.mill` file, which can be a problem sometimes.