Add a logger to Retry, and make the integration tests Retry effective (#4793)
This adds a `logger: String => Unit` argument to `Retry`, so that we can
know in the logs (CI logs say) if a retry did happen. Two helpers are
added to help users pass that new argument: `Retry.ctxLogger` that uses
`Task.ctx`, and `Retry.printStream` if one wants to log with
`System.err` for example.

Also, this moves the `Retry` of integration tests from
`UtestIntegrationTestSuite` to `IntegrationTestSuite#integrationTest`.
In the former location, it was wrapping a `Future`, which `Retry`
doesn't handle, so that the retry actually had no effect I think.

Lastly, this ignores `TimeoutException`s for example tests that have a
`ignoreErrorsOnCI` file (introduced in
https://github.com/com-lihaoyi/mill/pull/4751).