Rename `ZincWorker` to `JvmWorker` (#4834)
It no longer is exclusive to zinc, and nowadays also handles custom JVM
versions for compile/run/test as well. The new name is more generic to
reflect that, and probably easier to understand for the vast majority of
people who don't know what "Zinc" is

We could consider splitting out a `JvmWorker` from `ZincWorker`, but it
is likely the two will remain tightly bound together, e.g. to allow
https://github.com/com-lihaoyi/mill/issues/3982