Fix some recurrent CI errors (#4907)
This should fix different kinds of CI errors that we run into quite
often:
- some YAML / workflow definition errors
- junit tests failing on purpose in Mill test cases, that were reported
as CI errors in the workflow summary page by the action-junit-report
action
- GraalVM images, downloaded in the build-windows job, that went missing
in the examples / integration jobs (fixed by revalidation of some
`PathRef`s)
- Android linting errors printed on purpose by the Android tests, that
were matched anyway by problem matchers set up by the setup-android
action, and reported as errors on the workflow summary page
- timeout errors in the BSP tests (this one was my own mistake initially
😅)
- JVM version and JVM index-related errors: pinning the index version
isn't enough for the exact same JVM version (like `17.0.14`) to be
picked when passed only a major version (like `17`). If a lower version
is already in cache, coursier uses it, except if `JavaHome#update` is
set to true, which we do here. (This led to issues in the main repo CI,
that had older JVMs in cache, while forks' CI were usually fine)