Add ability to define `PackageExternalModule`s that can be called via just `pkg.path/` rather than `pkg.path.ObjectName/` (#4918)
This simplifies a common case where there is only one `ExternalModule`
in a package, e.g.

1. `./mill mill.scalalib.scalafmt.ScalafmtModule/` can now be called via
`./mill mill.scalalib.scalafmt/`
2. `./mill mill.kotlinlib.ktlint.KtlintModule/` can now be called via
`./mill mill.kotlinlib.ktlint/`
3. `./mill mill.javalib.palantirformat.PalantirFormatModule/` can now be
called via `./mill mill.javalib.palantirformat/`
4. `./mill mill.idea.GenIdea/` can now be called via `./mill mill.idea/`

This removes a huge source of verbosity from the Mill command line
interface, and basically fixes one of Mill's oldest issues
https://github.com/com-lihaoyi/mill/issues/299 by default. Referencing
external modules via their package path is sufficiently concise that it
basically removes the need for defining your own `def` external module
aliases, though you can always define your own aliases to make it more
concise still if you so desire

For cases where there is more than one `ExternalModule` in a single
`package`, they will need to be called fully qualified e.g. `./mill
mill.scalalib.PublishModule/`

`PackageExternalModule` is the spiritual equivalent to Scala `package
object` or Mill `TaskModule#defaultCommandName`, in that it allows you
to take a `package` in the hierarchy and designate one direct child as
the "blessed" entrypoint for that `package`.

The trailing `/` in the selector is still a bit annoying and weird, and
I might try making it optional a follow up PR.

## Implementation

The "blessed" `ExternalModule` in a package is specified by defining a 

```scala
object PackageExternalModule extends mill.define.PackageExternalModule(MyExternalModule)
```

Where `MyExternalModule` is the actual `ExternalModule` in question. The
`Resolve` logic when given a `foo.bar.baz/` selector first looks for the
class `foo.bar.baz.PackageExternalModule$` before looking for the class
`foo.bar.baz$`.

This adds a tiny amount of ambiguity in what a selector `foo.bar.baz`
means, but in practice I don't expect it to be a problem. I expect such
collisions between normal `ExternalModule` names and
`PackageExternalModule` names to happen basically never, so from a
practical perspective there will only ever be one obvious target for
each selector

I chose the indirection of defining a separate `PackageExternalModule`
object over renaming all the various external modules to
`PackageExernalModule`. Renaming them is possible, but then we lose the
meaningful name when referencing them in code, and breaks backwards
compatibility with the fully `foo.bar.baz.MyExternalModule/` selector
being passed at the command line.

Using an `object PackageExternalModule` as a marker lets us avoid doing
any sort of classpath scanning, which would be required if we wanted to
use an `@Annotation` or marker-trait for designating the primary
`ExternalModule` for a `package`. It just adds one additional
`ClassLoader#loadClass` check, which is much cheaper than scanning the
classpath with `classgraph` to try and find classes with particular
annotations

## Testing

Added a unit test case to cover this new resolution behavior, and
updated all the eligible example tests to use the shorter selector.

This could be backported to 0.12.x since it is backwards compatible