Rename `ivyDeps` to `mvnDeps` (#4930)
Fixes https://github.com/com-lihaoyi/mill/issues/4045

This PR replaces the `ivyDeps` nomenclature throughout Mill with
`mvnDeps`. This is a more accurate name, and the old name has long been
both confusing and inaccurate since Mill stopped using Ivy as the
dependency resolver many years ago.

An alternative I considered was `libraryDeps`: this has the upside of
being more consistent with SBT, but has the downside of being very long
as a string interpolator `library"..."`. And we can't easily get rid of
the string interpolator due to how type inference works, as implicit
conversions do not apply to the elements of a `Seq(...)` when the `Seq`
itself is being implicitly converted to `Task[Seq]`

We considered `jvmDeps`, but that's a bit confusing for things like
Scala.js/Scala-Native/Kotlin.js/etc. that aren't JVM

For now I left `import $ivy` in place, but I'm hoping to replace that
entirely as part of https://github.com/com-lihaoyi/mill/issues/3062. No
need to both renaming it now when it'll hopefully be gone soon anyway