Android: Hilt support with tranform ASM and todo external app sample build (#4759)
# Summary

~~This PR is for encouraging discussion on how to proceed with Hilt
support on Android. It composes of 2 main things:~~

~~1. How far I could go for Hilt support in Mill with current provided
tools~~
~~2. How Gradle does it and what's missing from Mill to achieve the same
result~~

This PR is now ready for review.

Covers part of https://github.com/com-lihaoyi/mill/issues/3868 (Hilt)

# How far this PR goes

With Kotlin KSP support added in
https://github.com/com-lihaoyi/mill/pull/4557 and
https://github.com/com-lihaoyi/mill/pull/4485 generating the Hilt source
files is relatively easy.

Java files are generated and I'm demonstrating in the PR how each
compilation step (Kotlin KSP, Kotlin compose, Java compilation and Java
annotation processing) are generated, following the exact steps gradle
is doing.

So in the end we have:

1. Kotlin compiled sources
2. Java sources generated from the KSP and Hilt compiler
4. Java compiled sources (including the _ComposedTreeDeps - an
aggregation of all dependencies that hilt does)
5. Transforms ASM for the dependency injection step in compile time
(only for main sources)
 

![image](https://github.com/user-attachments/assets/5949b73a-af4f-47e5-8e08-5a716bce779a)

Debug and test sources are not covered due to the amount of changes,
I've decided to add them in a future PR.

## The Transform ASM step

There is one step, that is done via a gradle plugin
(android-hilt-gradle-plugin) that is not straightforward to reuse. This
is what makes dependency injection possible.

In the demonstrated example, TodoActivity's constructor should look like
this after the final stage of compilation:

```
.method public constructor <init>()V
    .registers 1

    .line 29
    invoke-direct {p0}, Lcom/example/android/architecture/blueprints/todoapp/Hilt_TodoActivity;-><init>()V

    return-void
.end method
```

instead of the (normally) compiled

```
.method public constructor <init>()V
    .registers 1

    .line 29
    invoke-direct {p0}, Landroidx/activity/ComponentActivity;-><init>()V

    .line 28
    return-void
.end method
```

# How gradle does it

Gradle does the following steps on the release task (assembleRelease):

0. Fetch dependencies and jetify them (this I've skipped as it shouldn't
be needed with new android API and dependency versions)
1. Run KSP (kotlin with symbol processing plugin) which runs Dagger and
Hilt source generation, here the sources are written
2. Run Kotlin with compose plugin
3. Run java compilation on generated sources
4. Run java compilation on the component tree classes (there should just
be 2) with hilt-compiler as the annotation processor
6. Run transform ASM which achieves what is described in critical step

# What is needed

In my opinion, to achieve Hilt support for Mill (and have the android
architecture examples working) we need to implement Mill's version of
[AndroidEntryPointClassVisitor](https://github.com/google/dagger/blob/b3d3443e3581b8530cd85929614a1765cd37b12c/java/dagger/hilt/android/plugin/main/src/main/kotlin/dagger/hilt/android/plugin/transform/AndroidEntryPointClassVisitor.kt#L122)

With 2 instrumentation related dependencies, one specific for android
(`com.android.build.api.instrumentation`) and (`org.objectweb.asm`)
which is already in Mill if I'm not mistaken!


I would like your opinions on how to proceed!

## TL;DR 

Hilt should be possible after we implement or somehow reuse
android-hilt-gradle-plugin's transform ASM step to achieve the
dependency injection step for Activities


## Update

The transform asm step, is done in
`mill.main.android.hilt.AndroidHiltTransformAsm` which is provided as an
external program, that takes a directory to scan for classes and
re-write them to a destination directory.

The rewrite happens only if a class `isInstrumentable` , i.e. it
contains any of the following annotations:

```
        descriptor == "Ldagger/hilt/android/AndroidEntryPoint;" ||
        descriptor == "Ldagger/hilt/android/HiltAndroidApp;"
```

which is similar to [hilt's gradle
plugin](https://github.com/google/dagger/blob/b3d3443e3581b8530cd85929614a1765cd37b12c/java/dagger/hilt/android/plugin/main/src/main/kotlin/dagger/hilt/android/plugin/transform/AndroidEntryPointClassVisitor.kt#L122):

```
  override fun isInstrumentable(classData: ClassData) =
      classData.classAnnotations.any { ANDROID_ENTRY_POINT_ANNOTATIONS.contains(it) }
  }
  ```
  
## Other changes

I've introduced androidApplicationNamespace and androidApplicationId , but probably this needs a cleanup (we need a user defined package name to add to the android manifest). Gradle accepts these configs as:

```
android {
    namespace = "com.example.android.architecture.blueprints.todoapp"
    compileSdk = libs.versions.compileSdk.get().toInt()

    defaultConfig {
applicationId = "com.example.android.architecture.blueprints.main"
```

and since I have not used the application id in this PR after all, I can remove it until its need is more evident 


I've also changed the androidInstall to a command, firstly it can be run now more than once (instead of being cached) and in the future it can also accept parameters (e.g. to target a specific emulator)

EDIT: Application Id can be used and altered in build variants, we are gonna look into this soon, whether it stays or it's added later is not an issue for me

---------

Co-authored-by: Li Haoyi <haoyi.sg@gmail.com>