Add `Task.fail` API to conveniently fail a task (#4964)
Currently, to programmatically fail a task we can either throw an
exception or return a `Result.Failure`, but both ways have downsides.

Throwing an exception is easy but seems unintentional from the outside.
Instead of given a nice error message, we see a exception message and a
stack trace.
So, the nicer way is to return a `Result.Failure` with a nice message.
But if we want to return a `Result.Failure`, we also must wrap the
intended result in a `Result.Success`,
which is inonvenient.

By using the new `Task.fail`, we throw an dedicated exception type which
is autmatically converted into a `Result.Failure`,
but doesn't forces us to manually wrap any result into a
`Result.Success`.

Pull request: https://github.com/com-lihaoyi/mill/pull/4964