Android/cleanup app release install (#4996)
## Summary

- **better organisation** between androidmodule and androidappmodule
(separate properly what things are used in apps and things that are
shared between lib building and app building)
- R8 and D8 **dex generation** depending on build type settings
- **Instrumented tests** now use the android application namespace and
id and adding a "variant" indicator
- introduced a lite version of **build type settings** with main
motivation to control r8 and d8 but also serve as a scaffolding for
build variants
- **Android manifest** has less hardcoded parameters (uses-sdk and
instrumentation tags are now generated)
- Minor change into **testTask** to keep it consistent with the rest of
modules(Task.Anon) and thus the test-report is now in `testForked.dest`

## R8 and Proguard

Followed similar configuration style with gradle


![image](https://github.com/user-attachments/assets/c2c99135-6381-4c91-9ce2-ac5e666f5a33)

Although there's still work to be done in that area, if minification is
enabled, we use the R8 tool to generate the dex otherwise the d8 (which
also uses proguard for now) . This needs further work and clarification
as AGP seems to be using R8 by default when minification is enabled,
otherwise d8 is used.

Proguard default files are also fetched (as an opt-in setting) from the
Android SDK directories

## Build settings

I've introduced a first very minimal step for doing build variants,
mainly out of necessity to manage R8 and D8 variations better, but this
will be expanded once we need to fully support build variants.

The settings so far concerned the R8 use case at hand and it's a case
class that has the parameters to feed proguard files and flags into R8.

## R8 enhancements

Extending the work from https://github.com/com-lihaoyi/mill/pull/4892 ,
I've made some usability enhancements and enriched the examples to
demonstrate how instrumented tests can be used alongside R8 and release
builds.

I think R8 now can be fairly usable to generate minified and optimised
release builds without getting in the middle of development (e.g.
running tests and testing in the emulator) although more work needs to
be done towards that area, but we'll follow up after we do more
exploratory usability testing!

## Future work

Following this work, we can move to complete the work on architecture
samples by testing instrumented tests, unit tests and later on debug
variants.

The cleanup also reduced a lot of duplication between `AndroidAppModule`
and `AndroidModule` so it will make our development efforts a bit easier
to add more features!

More discovery work will follow to align the instrumented test apk
installation more in line with AGP