Remove requirement to `extends RootModule` (#5008)
Fixes https://github.com/com-lihaoyi/mill/issues/4971

With this PR, any top-level `object package` in a `build.mill` or
`package.mill` file has the `extends RootModule` or `extends
SubfolderModule` automatically inserted.

Before this PR we were already over-writing the `extends RootModule`
with the correct type, so this just makes it insert the correct type
even when `extends RootModule` is not provided (although we still
require that it extends some kind of `mill.Module`).

If there is an existing `extends RootModule`, we accept that for now to
ease in the migration