Fix selective execution when multiple changes are made to one module under --watch (main branch) (#5033)
Forward port to `main` of https://github.com/com-lihaoyi/mill/pull/5032

If you have two modules and run `--watch
{moduleA.fooCommand,moduleB.fooCommand}`, mill will run
`moduleB.fooCommand` every 2nd change to `moduleA`, even if there is no
changes to `moduleB`.

The problem stems from the fact that when we save the changed inputs in
selective execution, we did not store the inputs from `moduleB` and when
selective execution runs again (the 2nd time), it thinks that
*everything* in `moduleB` changed.

This fix correctly saves the state of all modules when storing the
selective execution state.

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>