Restrict `os.read` within a Task to only allow reading from upstream task `PathRef`s (#5037)
Fixes https://github.com/com-lihaoyi/mill/issues/5023, the last bullet
of https://github.com/com-lihaoyi/mill/issues/3746 we had punted on
earlier

With this PR:

1. A normal `Task` can only read from sub-paths of `PathRef`s returned
by the upstream tasks it depends on.
2. Like the write restrictions implemented earlier, we do not restrict
`Command`s (since those are expected to do arbitrary side-effecting
things),
3. For reads we are also lenient for `InputImpl`s and their sub-types
`SourceImpl`s and `SourcesImpl`s (since those are the roots of the build
graph and so are expected to read from arbitrary files)

To implement this, we need some way to traverse the return value of
upstream tasks. This is done by instrumenting the
`PathRef.jsonFormatter` that is called during the recursive
serialization/deserialization of task return values during cache saves
and loads. There is no other place for us to put this logic unless we
add a second typeclass that task return values would have to implement,
so this hackery piggy-backing on JSON serialization and de-serialization
will have to do for now

Added a single example test case to make sure this catches things,
relying on the bulk of the existing test suite to ensure this doesn't
break things. Also had to fix a bunch of existing violations of this
rule, mostly by adjusting the code to use the proper pattern, but for
some violations I just grandfathered them in for now. Definitely seems
like a good guardrail to have given how often we screw it up ourselves
in Mill's own repo

For now this only limits reads during the _execution_ phase, and does
not limit reads during the _resolution_ phase. That can come in a follow
up