Enforce classloader cleanup in unit tests (#5079)
Part of the effort to solve
https://github.com/com-lihaoyi/mill/issues/4522 by eliminating
classloader leaks

We record the classpath of any `Jvm.createClassLoader` and remove the
record every time `.close()` is called, and assert that there are no
created-but-not-closed classloaders in `UnitTester.<init>` or
`UnitTester#close`.

To make the existing tests pass, we needed to become much more
disciplined about using `UnitTester#scoped` rather than assigning `val
eval = new UnitTester` since the latter leaks the classloader

A lot of existing code uses hacky, leaky classloader caches in their
workers: playlib, scalapblib, etc. For now I replaced it all with the
most naive `Jvm.withClassLoader(){}` blocks to ensure things won't leak,
but we can go back and change them to `CachedFactories` in future if we
want to claw back that performance

There are some code paths that use raw `java.net.URLClassLoader`s that
our instrumentation is unable to track; left most of those as is, and we
can fix those in a follow up

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>