Set config to avoid CodeCache full problems (#5090)
Solves https://github.com/com-lihaoyi/mill/issues/4522. As far as I can
tell, the problem is simply high codecache memory usage due to the
different versions of the Scala compiler we have running around in
classloaders, but we're not actually leaking anything (at least since
https://github.com/com-lihaoyi/mill/pull/5079 and
https://github.com/com-lihaoyi/mill/pull/5080 and
https://github.com/com-lihaoyi/mill/pull/5082). So we just bump up the
codecache memory limit to make the problem go away

Also moved some scattered config files to the `//|` YAML header. This
needed fixes in `mill.bat`'s version header parsing to work.
`.mill-jvm-version` cannot be moved to the yaml header due to
`ClassCastException`, which I fixed in the code but will need a
re-bootstrapping to make use of in Mill's own build