Set context class loader when running tasks
It seems that's necessary on Windows for coursier, when it uses
a ServiceLoader to load some native utilities. ServiceLoader uses
the context class loader to look for services, and having it use
that loader addresses errors like
```
java.util.ServiceConfigurationError: coursier.jniutils.NativeApi: coursier.jniutils.DefaultNativeApi not a subtype
```

Plus it seems it's a good practice to set that loader to a sensible
value when class loading and running external code.