Split launcher/daemon terminology from client/server (#5135)
We always had two clients and two servers: previously
`mill.main.client`/`mill.runner.client` and
`mill.main.server`/`mill.runner.server`, and when you saw `client` or
`server` it was never clear which one you were referring to.
Furthermore, `client`/`server` was always very confusing since it brings
to mind HTTP clients/server or network client/servers which Mill is not.

In `main` with #5062 we started using `client`/`server` to refer to the
underlying protocol, while `launcher`/`daemon` referred to the
high-level JVM processes. These are much better names for the two
concepts, and avoid conflicting with the `client`/`server` terms used
for the underlying protocol components

This PR standardizes our terminology to use `launcher`/`daemon`
throughout the codebase, with the `.fork`/`.server` names in our
integration tests renamed to `.nodaemon`/`.daemon`