Consolidate build-global APIs in `mill.define.BuildCtx` (#5142)
This PR standardizes the location of APIs that are global to the build
on `mill.define.BuildCtx`, meant to be analogous to `EvalCtx` and
`ModuleCtx`. Previously they were scattered on
`mill.define.WorkspaceApi`,`build.interp`, and some direct calls to
`os.checker.withValue`. Now, they all are consolidated under `BuildCtx`
where they can be discovered and documented.

Since the `watch` lists aren't conveniently bundled together with
`RootModule` anymore, we cannot rely on `RootModuleApi` when passing
them around `MillBuildBootstrap`, so I wrapped them in a `BuildFileApi`
object to pass them around conveniently. This subsumes the previous
`wrapper_object_getter` that we code-gened